Fixed a bunch of compiler warnings

+++ gtsam/base/Matrix.cpp
@@ -682,7 +682,7 @@ void svd(const Matrix& A, Matrix& U, Vector& S, Matrix& V) {
 boost::tuple<int, double, Vector> DLT(const Matrix& A, double rank_tol) {
 
   // Check size of A
-  int n = A.rows(), p = A.cols(), m = min(n,p);
+  size_t n = A.rows(), p = A.cols(), m = min(n,p);
 
   // Do SVD on A
   Eigen::JacobiSVD<Matrix> svd(A, Eigen::ComputeFullV);
@@ -690,20 +690,20 @@ boost::tuple<int, double, Vector> DLT(const Matrix& A, double rank_tol) {
   Matrix V = svd.matrixV();
 
   // Find rank
-  int rank = 0;
-  for (int j = 0; j < m; j++)
+  size_t rank = 0;
+  for (size_t j = 0; j < m; j++)
     if (s(j) > rank_tol) rank++;
 
   // Return rank, error, and corresponding column of V
   double error = m<p ? 0 : s(m-1);
-  return boost::tuple<int, double, Vector>(rank, error, Vector(column(V, p-1)));
+  return boost::tuple<int, double, Vector>((int)rank, error, Vector(column(V, p-1)));
 }
 
 /* ************************************************************************* */
 Matrix expm(const Matrix& A, size_t K) {
   Matrix E = eye(A.rows()), A_k = eye(A.rows());
   for(size_t k=1;k<=K;k++) {
-    A_k = A_k*A/k;
+    A_k = A_k*A/double(k);
     E = E + A_k;
   }
   return E;
@@ -711,7 +711,7 @@ Matrix expm(const Matrix& A, size_t K) {
 
 /* ************************************************************************* */
 Matrix Cayley(const Matrix& A) {
-  int n = A.cols();
+  size_t n = A.cols();
   assert(A.rows() == n);
 
   // original
